✅ Add garbage collector info tests
