+++ .github/workflows/kernel-coverage.yml
@@ -120,11 +120,25 @@ jobs:
       step_timeout_minutes: 5
       coverage: true
 
+  # ── GarbageCollector ──────────────────────────────────────────────────────────
+  garbagecollector-tests:
+    needs: setup
+    if: always() && needs.setup.result == 'success'
+    uses: ./.github/workflows/kernel-test-run.yml
+    with:
+      suite_name: GarbageCollector
+      suite_id: garbagecollector
+      kernel_path: tests/Kernels/Cosmos.Kernel.Tests.GarbageCollector
+      timeout_x64: 120
+      timeout_arm64: 180
+      step_timeout_minutes: 10
+      coverage: true
+
   # ── Coverage Report ──────────────────────────────────────────────────────────
   coverage-report:
     name: Coverage Report
     runs-on: ubuntu-latest
-    needs: [helloworld-tests, memory-tests, typecasting-tests, timer-tests, network-tests, threading-tests, graphic-tests]
+    needs: [helloworld-tests, memory-tests, typecasting-tests, timer-tests, network-tests, threading-tests, graphic-tests, garbagecollector-tests]
     if: always() && (github.event_name == 'workflow_dispatch' || contains(github.event.pull_request.labels.*.name, 'run-coverage'))
     permissions:
       pull-requests: write
