+++ .github/workflows/kernel-tests.yml
@@ -203,11 +203,34 @@ jobs:
       event_name: ${{ github.event_name }}
       head_ref: ${{ github.head_ref || github.ref_name }}
 
+  # ── GarbageCollector ──────────────────────────────────────────────────────────
+  garbagecollector-tests:
+    needs: setup
+    uses: ./.github/workflows/kernel-test-run.yml
+    with:
+      suite_name: GarbageCollector
+      suite_id: garbagecollector
+      kernel_path: tests/Kernels/Cosmos.Kernel.Tests.GarbageCollector
+      timeout_x64: 120
+      timeout_arm64: 180
+      step_timeout_minutes: 10
+
+  garbagecollector-results:
+    needs: garbagecollector-tests
+    if: always()
+    uses: ./.github/workflows/kernel-test-results.yml
+    with:
+      suite_name: GarbageCollector
+      suite_id: garbagecollector
+      kernel_path: tests/Kernels/Cosmos.Kernel.Tests.GarbageCollector
+      event_name: ${{ github.event_name }}
+      head_ref: ${{ github.head_ref || github.ref_name }}
+
   # ── Summary ───────────────────────────────────────────────────────────────────
   test-summary:
     name: Test Summary
     runs-on: ubuntu-latest
-    needs: [ helloworld-results, memory-results, typecasting-results, timer-results, network-results, threading-results, graphic-results ]
+    needs: [ helloworld-results, memory-results, typecasting-results, timer-results, network-results, threading-results, graphic-results, garbagecollector-results ]
     if: always()
     steps:
       - name: 📊 Generate Summary
@@ -227,13 +250,14 @@ jobs:
             fi
           }
 
-          add_row "HelloWorld"  "${{ needs.helloworld-tests.result }}"
-          add_row "Memory"      "${{ needs.memory-tests.result }}"
-          add_row "TypeCasting" "${{ needs.typecasting-tests.result }}"
-          add_row "Timer"       "${{ needs.timer-tests.result }}"
-          add_row "Network"     "${{ needs.network-tests.result }}"
-          add_row "Threading"   "${{ needs.threading-tests.result }}"
-          add_row "Graphic"     "${{ needs.graphic-tests.result }}"
+          add_row "HelloWorld"       "${{ needs.helloworld-tests.result }}"
+          add_row "Memory"           "${{ needs.memory-tests.result }}"
+          add_row "TypeCasting"      "${{ needs.typecasting-tests.result }}"
+          add_row "Timer"            "${{ needs.timer-tests.result }}"
+          add_row "Network"          "${{ needs.network-tests.result }}"
+          add_row "Threading"        "${{ needs.threading-tests.result }}"
+          add_row "Graphic"          "${{ needs.graphic-tests.result }}"
+          add_row "GarbageCollector" "${{ needs.garbagecollector-tests.result }}"
 
           echo "" >> $GITHUB_STEP_SUMMARY
           echo "**Run ID:** ${{ github.run_id }}" >> $GITHUB_STEP_SUMMARY
