+++ tests/Kernels/Cosmos.Kernel.Tests.GarbageCollector/Bootloader/limine.conf
@@ -1,0 +1,10 @@
+# Timeout in seconds that Limine will use before automatically booting.
+timeout: 0
+
+# The entry name that will be displayed in the boot menu.
+/Limine Template
+    # We use the Limine boot protocol.
+    protocol: limine
+
+    # Path to the kernel to boot. boot():/ represents the partition on which limine.conf is located.
+    path: boot():/boot/Cosmos.Kernel.Tests.GarbageCollector.elf
